refactor: reorganize project structure
