+++ src/features/Home/components/AboutSection.tsx
@@ -1,5 +1,5 @@
 import { motion } from "framer-motion";
-import styles from "../../../styles/AboutSection.module.css";
+import styles from "./AboutSection.module.css";
 import { profile } from "../../../data/profile";
 
 const AboutSection = () => {
