+++ src/features/Home/components/HeroSection.tsx
@@ -1,5 +1,5 @@
 import { motion } from "framer-motion";
-import styles from "../../../styles/HeroSection.module.css";
+import styles from "./HeroSection.module.css";
 import { profile } from "../../../data/profile";
 
 const HeroSection = () => {
